+++ SurvivalGame/Assets/Crafting/ItemCraftBlueprint.cs
@@ -1,4 +1,6 @@
 using UnityEngine;
+using System.Collections.Generic;
+using System.Linq;
 
 [CreateAssetMenu(fileName = "NewItemCraftBlueprint", menuName = "Game/ItemCraftBlueprint")]
 public class ItemCraftBlueprint : CraftBlueprint
@@ -6,9 +8,35 @@ public class ItemCraftBlueprint : CraftBlueprint
     public ItemData outputItem;
     public int outputQuantity;
 
+    public CraftStationType requiredCraftStationType = CraftStationType.None;
+    private float craftStationRequiredDistance = 3.5f;
+
     public bool CanCraft(InventorySystem inventory)
     {
-        // First simple check is to check whenever the required items are available
+        if (!IsCraftStationAvailable())
+            return false;
+
+        if (!HasRequiredItems(inventory))
+            return false;
+
+        if (WillCauseOverweight(inventory))
+            return false;
+
+        return HasSpaceForOutput(inventory);
+    }
+
+    private bool IsCraftStationAvailable()
+    {
+        bool craftStationAvailable = CheckRequiredCraftingStationAvailability();
+        if (!craftStationAvailable)
+        {
+            Debug.Log("[ItemCraftBlueprint] Can't craft item because required craftStation is not found in range.");
+        }
+        return craftStationAvailable;
+    }
+
+    private bool HasRequiredItems(InventorySystem inventory)
+    {
         foreach (ItemRequirement requirement in requirements)
         {
             if (!inventory.isItemAvailable(new ItemInstance(requirement.item, requirement.quantity)))
@@ -17,31 +45,60 @@ public class ItemCraftBlueprint : CraftBlueprint
                 return false;
             }
         }
+        return true;
+    }
 
-        // Now i need to check if there is space for the output
+    private bool HasSpaceForOutput(InventorySystem inventory)
+    {
         if (inventory.GetFreeSlots() > 0)
-        {
             return true; // Easy case, a slot is empty
-        }
 
-        // Medium case, maybe there is a slot with already the output item
         if (inventory.IsThereASlotWithItem(outputItem))
-        {
-            return true;
-        }
+            return true; // Medium case, slot with already the output item
 
-        // Hard case, i need to check if a slot would be freed after removal of required input
-        foreach (ItemRequirement requirement in requirements)
-        {
-            if (inventory.willRemovalFreeASlot(new ItemInstance(requirement.item, requirement.quantity)))
-            {
-                return true;
-            }
-        }
+        if (requirements.Any(req => inventory.willRemovalFreeASlot(new ItemInstance(req.item, req.quantity))))
+            return true; // Hard case, check if removal of required items frees a slot
 
         Debug.Log("[ItemCraftBlueprint] Can't craft item because no free slot available after using the required items");
         return false;
+    }
 
+    private bool WillCauseOverweight(InventorySystem inventory)
+    {
+        float addedWeight = outputItem.weight * outputQuantity;
+        float removedWeight = requirements.Sum(req => req.item.weight * req.quantity);
+        float deltaWeight = addedWeight - removedWeight;
+
+        if (deltaWeight > 0 && inventory.WillBeOverWeight(deltaWeight))
+        {
+            Debug.Log("[ItemCraftBlueprint] Can't craft item since it would cause overweight");
+            return false;
+        }
+        return true;
+    }
+
+    private bool CheckRequiredCraftingStationAvailability()
+    {
+        if (requiredCraftStationType != CraftStationType.None)
+        {
+            GameManager gm = GameManager.Instance;
+            Vector3 playerPos = gm.getPlayer().transform.position;
+            GridManager gridManager = gm.getTerrainGridManager();
+
+            List<GameObject> craftStations = gridManager.GetAllObjectsOfTypeOnTilesByRadius(playerPos, craftStationRequiredDistance, typeof(CraftStationBuillding));
+            foreach (GameObject station in craftStations)
+            {
+                if (station.GetComponent<CraftStationBuillding>().IsCraftStationOfType(requiredCraftStationType))
+                {
+                    return true;
+                }
+            }
+            return false;
+        }
+        else
+        {
+            return true;
+        }
     }
 
     public bool Craft(InventorySystem inventory)