+++ SurvivalGame/Assets/Editor/ItemCraftBlueprintDebug.cs
@@ -14,15 +14,8 @@ public class ItemCraftBlueprintDebug : Editor
         {
             InventorySystem inventory = GameManager.Instance.getInventorySystem();
 
-            bool canCraft = blueprint.CanCraft(inventory);
-            bool result = false;
-            if (canCraft)
-            {
-                result = blueprint.Craft(inventory);
-            }
+            blueprint.Craft(inventory);
 
-            string resultMessage = canCraft ? (result ? "Crafting successful!" : "Crafting failed!") : "Cannot craft due to insufficient resources or space.";
-            Debug.Log(resultMessage);
         }
     }
 }